It very nearly almost compiles
Only one minor issue with implementations of func_192166_a that we need
to look into.

All API patches were already finished prior but I've moved them to a
separate folder so we can verify compile without shuffling them
constantly. As we add the implementations we can put them back in the
proper directory.