Upgrade to PaperGradle 2.3.12
This update brings Paperclip support to PaperMCP. This also fixes a
patch that was broken for some reason, and fixes an issue with the
reobfuscate task where Spigot moved an internal Minecraft field, causing
errors at runtime trying to access a now non-existent field. This was
fixed with the addition of the member-moves.txt file, along with the new
patchSrgs task which runs before (and provides the input for) the
reobfuscate task.

Paper-Server wasn't originally setup for the build task to depend on the
shadowJar task, which broken an expectation in the reobfuscate task in
PaperGradle, which this commit fixes as well.