Add a flag, "disable-ghost-block-fix", to allow for turning off the GhostBlockFix https://github.com/PaperMC/Paper/pull/2396
The reason for this is that this fix re-introduces
this bug https://bugs.mojang.com/browse/MC-156013

This gives server owners a way to opt out of the GhostBlockFix.
It defaults to false, which is keeping the GhostBlockFix.

From: Zean Rivera <zean@zean.org>
Date: Sat, 7 Sep 2019 03:11:05 -0700
Subject: [PATCH] Add a flag, "disable-ghost-block-fix", to allow for turning off the GhostBlockFix https://github.com/PaperMC/Paper/pull/2396

The reason for this is that this fix re-introduces
this bug https://bugs.mojang.com/browse/MC-156013

This gives server owners a way to opt out of the GhostBlockFix.
It defaults to false, which is keeping the GhostBlockFix.